Add WorldID Identity Manager for EIP-4844
Implement WorldID Identity Manager in its 3rd implementation.
The new implementation contains one main method - registerIdentities.
It's the EIP-4844 equivalent of the existing registerIdentities
from the V1 implementation.

The new registerIdentities calls verifyProof that is also different from
its classic counterpart. The new verifyProof comes from the
EIP-4844-ready InsertionTreeVerifier164844 contract, that was generated
from the semaphore-mtb's InsertionCircuit modified to accommodae
EIP-4844 features.

Since the new verifyProof function has a different prototype than the
previous one, the ITreeVerifier interface has been updated with the new
prototype. This change results in lots of boilerplate changes as all
contracts implementing this interface must contain a dummy implementation
of the new function, even if it does not make practical sense.

Add tests for the new feature. Tests are modelled by the existing
registration tests. The new contract is basically a copy of the old one,
with the main difference of calling the new registerIdentities
implementation.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>